+++ CHANGELOG.md
@@ -47,8 +47,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added config option `picker.tag_mappings`, analogous to `picker.note_mappings`.
 - Added `log` field to `obsidian.Client` for easier access to the logger.
-- Added ability to follow basic anchor links. Nested anchor links are not yet supported.
+- Added ability to follow anchor links.
 - Added completion support for header anchor links. Note that this requires you to update your `wiki_link_func` and `markdown_link_func` in your config to handle anchors. See the configuration example in the README.
+- You can set `wiki_link_func` to a one of the following strings to use a builtin function:
+  - `"use_alias_only"`, e.g. `[[Foo Bar]]`
+  - `"prepend_note_id"`, e.g. `[[foo-bar|Foo Bar]]`
+  - `"prepend_note_path"`, e.g. `[[foo-bar.md|Foo Bar]]`
+  - `"use_path_only"`, e.g. `[[foo-bar.md]]`
 
 ### Changed